services/serverAvatar/pyproject.toml
[project]
name = "avatareleven"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "websockets>=15.0.1",
]
